Extract SwiftParser to separate file
Swift class is using `fs` and `path` and `chokidar` that are file system
dependent. In order to use swift parser inside browser (with webpack) we
need to extract SwiftParser to work with data streams instead of files.
