+++ lib/swift.js
@@ -1,13 +1,11 @@
 const fs = require('fs');
 const path = require('path');
-const finParser = require('./finParser.js');
-const mtParser = require('./mtParser.js');
-const block1Parser = require('./block1Parser');
-const block2Parser = require('./block2Parser');
 const messageGenerator = require('./messageGenerator');
-const chokidar = require('chokidar');
 const { Logger, mkDirByPathSync } = require('./utils');
 const moment = require('moment');
+const patterns = require('./metadata/patterns.json');
+const chokidar = require('chokidar');
+const SwiftParser = require('./swiftParser');
 
 const logger = new Logger();
 
@@ -17,8 +15,9 @@ class Swift {
       this.fieldPatterns =
         JSON.parse(fs.readFileSync(path.resolve(process.cwd(), options.fieldPatterns)));
     } else {
-      this.fieldPatterns = JSON.parse(fs.readFileSync(`${__dirname}/metadata/patterns.json`));
+      this.fieldPatterns = JSON.parse(patterns);
     }
+    this.parser = new SwiftParser(this.fieldPatterns);
     this.inputFolder = options.in ? path.resolve(process.cwd(), options.in) : path.resolve(process.cwd(), './in');
     this.outputFolder =
       options.out ? path.resolve(process.cwd(), options.out) : path.resolve(process.cwd(), './out');
@@ -189,36 +188,7 @@ class Swift {
   }
 
   parse(swiftMessage) {
-    // eslint-disable-next-line no-param-reassign
-    swiftMessage = swiftMessage.replace(/\r\n/g, '\n');
-    const ast = finParser.parse(swiftMessage);
-
-    const humanizeBlocks = (blocks) => {
-      Object.keys(blocks).forEach((name) => {
-        switch (name) {
-          case 'block1':
-            // eslint-disable-next-line no-param-reassign
-            blocks.block1 = block1Parser.parse(blocks.block1.content[0]);
-            break;
-          case 'block2':
-            // eslint-disable-next-line no-param-reassign
-            blocks.block2 = block2Parser.parse(blocks.block2.content[0]);
-            break;
-          case 'extraBlocks':
-            // eslint-disable-next-line no-param-reassign
-            blocks.extraBlocks = humanizeBlocks(blocks.extraBlocks);
-            break;
-          default:
-            // eslint-disable-next-line no-param-reassign
-            blocks[name] = mtParser.parse(blocks[name], this.fieldPatterns);
-            break;
-        }
-      });
-      return blocks;
-    };
-    const res = humanizeBlocks(ast);
-
-    return res;
+    return this.parser.parse(swiftMessage);
   }
 
   // eslint-disable-next-line class-methods-use-this
